requirements.txt
PyQT6~=6.2.0
tables==3.6.1
acoular==21.05
numpy~=1.20.3
Pillow~=8.3.2
matplotlib~=3.4.3
pyqtgraph~=0.12.3
scipy~=1.7.1
PyAudio~=0.2.11
opencv-python~=4.5.3.56